# models/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "API to scan and indentify the malware urls"
  version: "1.0.0"
  title: "malware url detection service"
host: "localhost"
schemes:
- "http"
paths:
  /urlinfo/1/{hostname_and_port}/{original_path_and_query_string}:
    get:
      tags:
      - "urlinfo"
      summary: "Finding the malware url by hostname and port"
      description: "Returns the url info by hostname and port"
      operationId: "getURLInfo"
      produces:
      - "application/json"
      parameters:
      - name: "hostname_and_port"
        in: "path"
        description: "hostname and port of the url to return"
        required: true
        type: "string"
        x-exportParamName: "HostnameAndPort"
      - name: "original_path_and_query_string"
        in: "path"
        description: "original_path and query_string of the url to return"
        required: true
        type: "string"
        x-exportParamName: "OriginalPathAndQueryString"
      responses:
        200:
          description: "This is the malware url"
          schema:
            $ref: "#/definitions/ApiResp"
        204:
          description: "This is not a malware url"
          schema:
            $ref: "#/definitions/ApiResp"
  /urlupdate/1/:
    post:
      tags:
      - "urlinfo"
      summary: "Update the datastore with malware urls"
      description: "Update the datastore with malware urls"
      operationId: "updateURLInfo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "malware url that needs to be updated in the datastore"
        required: true
        schema:
          $ref: "#/definitions/SetUrl"
        x-exportParamName: "Body"
      responses:
        200:
          description: "Successful operation"
          schema:
            $ref: "#/definitions/ApiResp"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ApiResp"
definitions:
  SetUrl:
    type: "object"
    properties:
      URL:
        type: "string"
    example:
      URL: "URL"
  ApiResp:
    type: "object"
    properties:
      Status:
        type: "string"
      Code:
        type: "integer"
        format: "int"
      Message:
        type: "string"
    example:
      Status: "Status"
      Message: "Message"
      Code: 0