# models/api_resp.go
/*
 * malware url detection service
 *
 * API to scan and indentify the malware urls
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ApiResp struct {

	Status string `json:"Status,omitempty"`

	Code int32 `json:"Code,omitempty"`

	Message string `json:"Message,omitempty"`
}