# models/set_url.go
/*
 * malware url detection service
 *
 * API to scan and indentify the malware urls
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type SetUrl struct {

	URL string `json:"URL,omitempty"`
}